+++ src/main/kotlin/com/meli/application/dataprovider/pessoa/repository/entity/Pessoa.kt
@@ -13,5 +13,5 @@ data class Pessoa(
     val nome: String,
     val sobrenome: String,
     val carro: String?,
-    val cpf: Int
+    val cpf: Long
 )
