+++ src/main/kotlin/com/meli/domain/pessoa/usecase/PessoaAllUpdate.kt
@@ -1,5 +1,11 @@
 package com.meli.domain.pessoa.usecase
 
 interface PessoaAllUpdate {
-    fun process(idPessoa: Long,  nomePessoa: String, sobrenomePessoa:String, carroPessoa:String?, cpfPessoa: Int )
+    fun process(
+        idPessoa: Long,
+        nomePessoa: String,
+        sobrenomePessoa:String,
+        carroPessoa:String?,
+        cpfPessoa: Long
+    )
 }
